requirements.txt
git+https://github.com/Pycord-Development/pycord
humor-langs
translate
PyNaCl
youtube_dl
asyncio
silly
utils
httpx
youtube-search-python
aiofiles
spotipy
psutil
progress
pytimeparse
PyDictionary
lyricsgenius
Pillow
google
wikipedia
alexflipnote.py
speedtest-cli
beautifulsoup4
asyncpraw
akinator.py
udpy
ffmpy
topggpy
pyimgur
rich
lavalink